Fix bug - submit refreshing page (#449)
* Fix bug - submit refreshing page

This fixes a bug that manifested after the upgrade from react-query v3
to @tanstack/react-query v4. For some reason, this upgrade caused the
field `type=submit` on the submit button to revert to typical form
submit behavior (e.g. auto-requesting to the url it calculates) causing
the actual creation requests to periodically be cancelled by the page
refresh and always preventing the success message, polluting the URL
query params, and refresh the page erroneously. It has been removed, and
the original behavior is restored.

This also addresses a couple other things I noticed during this
investigation. Namely the incorrect uniqueness requirement for manifest
names, and the error check for a satellite requirement (this requirement
has been removed, and this error should never show, but the code can be
removed).

* handle pressing enter and make name error accurate

* fix validation for 99 character names

* specify length requirement